requirements.txt
streamlit>=1.32.2
opencv-python-headless>=4.9.0
ultralytics>=8.3.0
shapely>=2.0.3
numpy<2.0.0
Pillow>=10.2.0